# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "User Processing API",
        "description": "API para processamento de usuários",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/users/process": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Processa um usuário (validação, integração e cache)",
                "operationId": "4f08e71c729cad17e9434c95a659ad73",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cpf",
                                    "cep",
                                    "email"
                                ],
                                "properties": {
                                    "cpf": {
                                        "type": "string",
                                        "example": "12345678900"
                                    },
                                    "cep": {
                                        "type": "string",
                                        "example": "06454000"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "usuario@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso (cache hit ou processamento)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "ok"
                                        },
                                        "source": {
                                            "type": "string",
                                            "example": "api"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validação falhou",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The cep field is required."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Erro em APIs externas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "external_api_error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/users/{cpf}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Busca usuário por CPF (com cache)",
                "operationId": "41877e3f5a32fae39960a2211845a7cb",
                "parameters": [
                    {
                        "name": "cpf",
                        "in": "path",
                        "description": "CPF do usuário",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "12345678900"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "ok"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Users",
            "description": "Users"
        }
    ]
}